pclientd: move implementation into library code
This means that the binary is just a thin wrapper around reusable library code.
